Updated all scripts for publisher.jar url + suppressed proprietaty codes from examples
Tried to suppress the slicing info message in the ignoreWarnings.txt file but this does not work. So I suppressed the 3 Quest proprietary codes from the examples.
I also suppressed the FHIR-order-catalog.xml file, which does not belong to this github repository but to the  http://github.com/hl7/jIRA-Spec-Artifacts project.